remove envelope attachment
